prototype3/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
Flask-Session==0.5.0
google-generativeai==0.3.2
requests==2.31.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
